Use MPI_Comm_c2f to convert an MPI communicator to a Fortran integer.

+++ src/opsinputs/VarObsWriter.interface.F90
@@ -46,7 +46,7 @@ use oops_variables_mod
 implicit none
 integer(c_int), intent(inout)  :: c_self
 type(c_ptr), value, intent(in) :: c_conf
-integer(c_size_t), intent(in)  :: c_comm ! MPI communicator to be used by OPS
+integer, intent(in)            :: c_comm ! MPI communicator to be used by OPS
 integer(c_int), intent(in)     :: c_nchannels
 integer(c_int), intent(in)     :: c_channels(c_nchannels)
 type(c_ptr), intent(in), value :: c_varlist ! list of geovals variables to be requested
